Share Integration Tests Between Server Implementations
